Polish payment and account UI modules
