+++ apps/portal/src/features/account/account-management/index.ts
@@ -1,0 +1,2 @@
+export { AccountManagement } from "./account-management";
+export type { AddSignInOption, AddWalletOption } from "./account-management";
